dbMemberId field was appended

+++ contracts/pluto/Pluto.sol
@@ -1,15 +1,15 @@
 pragma solidity ^0.4.17;
 
-import '../zeppelin-solidity/ownership/Ownable.sol';
-import './PlutoWallet.sol';
+import "../zeppelin-solidity/ownership/Ownable.sol";
+import "./PlutoWallet.sol";
 
 
 contract Pluto is Ownable {
-    mapping (uint64 => address) public mWallets;
+    mapping (uint => address) public mWallets;
 
-    event WalletCreated(uint64 indexed memberId, address indexed walletAddr);
+    event WalletCreated(uint indexed memberId, address indexed walletAddr);
 
-    function createWallet(uint64 _memberId)
+    function createWallet(uint _memberId)
         onlyOwner
         returns (address walletAddr)
     {
