+++ contracts/pluto/PlutoWallet.sol
@@ -1,12 +1,12 @@
 pragma solidity ^0.4.17;
 
-import '../zeppelin-solidity/ownership/Ownable.sol';
+import "../zeppelin-solidity/ownership/Ownable.sol";
 
 
 contract PlutoWallet is Ownable {
-    uint64 public mMemberId;
+    uint public mMemberId;
 
-    function PlutoWallet(uint64 _memberId) {
+    function PlutoWallet(uint _memberId) {
         mMemberId = _memberId;
     }
 }
